+++ gcc/rust/expand/rust-macro-builtins.h
@@ -31,9 +31,9 @@ namespace Rust {
 // transcriber and extra info if necessary
 // then make a global map<string, BuiltinMacro>
 
-/**
- * All builtin macros possible
- */
+//
+// All builtin macros possible
+//
 enum class BuiltinMacro
 {
   Assert,
@@ -79,42 +79,43 @@ enum class BuiltinMacro
 tl::optional<BuiltinMacro>
 builtin_macro_from_string (const std::string &identifier);
 
-/**
- * This class provides a list of builtin macros implemented by the compiler.
- * The functions defined are called "builtin transcribers" in that they replace
- * the transcribing part of a macro definition.
- *
- * Like regular macro transcribers, they are responsible for building and
- * returning an AST fragment: basically a vector of AST nodes put together.
- *
- * Unlike regular declarative macros where each match arm has its own associated
- * transcriber, builtin transcribers are responsible for handling all match arms
- * of the macro. This means that you should take extra care when implementing a
- * builtin containing multiple match arms: You will probably need to do some
- * lookahead in order to determine which match arm the user intended to use.
- *
- * An example of this is the `assert!()` macro:
- *
- * ```
- *  macro_rules! assert {
- *	($cond:expr $(,)?) => {{ ... }};
- *	($cond : expr, $ ($arg : tt) +) = > {{ ... }};
- * }
- * ```
- *
- * If more tokens exist beyond the optional comma, they need to be handled as
- * a token-tree for a custom panic message.
- *
- * These builtin macros with empty transcribers are defined in the standard
- * library. They are marked with a special attribute, `#[rustc_builtin_macro]`.
- * When this attribute is present on a macro definition, the compiler should
- * look for an associated transcriber in the mappings. Meaning that you must
- * remember to insert your transcriber in the `builtin_macros` map of the
- *`Mappings`.
- *
- * This map is built as a static variable in the `insert_macro_def()` method
- * of the `Mappings` class.
- */
+//
+//    This class provides a list of builtin macros implemented by the compiler.
+//    The functions defined are called "builtin transcribers" in that they
+//    replace the transcribing part of a macro definition.
+//
+//    Like regular macro transcribers, they are responsible for building and
+//    returning an AST fragment: basically a vector of AST nodes put together.
+//
+//    Unlike regular declarative macros where each match arm has its own
+//    associated transcriber, builtin transcribers are responsible for handling
+//    all match arms of the macro. This means that you should take extra care
+//    when implementing a builtin containing multiple match arms: You will
+//    probably need to do some lookahead in order to determine which match arm
+//    the user intended to use.
+//
+//    An example of this is the `assert!()` macro:
+//
+//    ```
+//    macro_rules! assert {
+//    ($cond:expr $(,)?) => {{ ... }};
+//    ($cond : expr, $ ($arg : tt) +) = > {{ ... }};
+//    }
+//    ```
+//
+//    If more tokens exist beyond the optional comma, they need to be handled as
+//    a token-tree for a custom panic message.
+//
+//    These builtin macros with empty transcribers are defined in the standard
+//    library. They are marked with a special attribute,
+//    `#[rustc_builtin_macro]`. When this attribute is present on a macro
+//    definition, the compiler should look for an associated transcriber in the
+//    mappings. Meaning that you must remember to insert your transcriber in the
+//    `builtin_macros` map of the `Mappings`.
+//
+//    This map is built as a static variable in the `insert_macro_def()` method
+//    of the `Mappings` class.
+
 class MacroBuiltin
 {
 public:
